Switched save game name typing to use love.textinput()
When the user types the name of a savegame, the text now goes through
love.textinput() instead of love.keypressed(). The difference is that
LOVE interprets characters the way a text editor would: shift+a is A;
shift+2 is @. The other difference is that the string is encoded as
utf-8, which means that characters can be multiple bytes. Fortunately,
LOVE also provides a library to handle the utf-8 encoding.