Introduce restitution to collisions
Instead of manually calcualting linear impulses when things collide, use
the built-in restitution property of objects in Box2D. This means that
structures now bounce off each other even when no parts are destroyed,
but we lose the calucation we previously made to account for lost
kinetic energy due to deformation of destroyed parts. This calculation
was inaccurate anyway. We may add it back eventually.