Allow startup in headless mode (#230)
Headless mode in general doesn't work yet, but it's close. The unit tests that exit before the main game loop work. When I turn off love.draw(), the main menu works fine.

However, trying to load a scene crashes because it tries to make a camera. I will leave fixing that for a different commit.

* Import lume version 98847e7812cf28d3d64b289b03fad71dc704547d

https://github.com/rxi/lume/commit/98847e7812cf28d3d64b289b03fad71dc704547d

* Move all graphics calls to inside love.draw()

* Memoize image loading

* Factor our common parts of draw functions

* More cleanup (thanks lunatest)