Ignore unused function paramaters in luacheck
I don't like configuring a linter with project-specific details, but I
have grown to believe that this particular check is not very useful. We
have a lot of functions that follow a standard API. They don't use all
of the arguments they receive, but all functions that use the same API
have the same arguments.

This is similar to Go, where an unused variable is an error. But
unused function parameters are not considered unused variables.