Remove drawPolygon proxy function
For some reason I thought that if you put a function that returns
multiple values as an argument to another function, it would only take
the first return value and discard the rest, UNLESS the function call is
the only argument. That is not true at all. The all the return values
get passed as arguments in every case. That makes this workaround I did
unneccessary.