Change world's event system to take a closure instead of a table
The goal of this change is to put the responsibility of creating objects
on the code that wants to put a new object in the world. world.lua still
calls the code to create objects at the correct time and is still
responsible for inserting objects into world.objects. However, if there
is an error in the code that generates a "create" event, it now looks
like that code's fault instead of the world's fault. This makes it
easier to debug errors in create events.

Here is an example. Let's modify structure.lua so that it generates an
invalid structure. Instead of a valid location, we'll pass in a number.

	diff --git a/src/world/structure.lua b/src/world/structure.lua
	index 985ea8a..985f0fe 100644
	--- a/src/world/structure.lua
	+++ b/src/world/structure.lua
	@@ -343,7 +343,7 @@ function Structure:disconnectPart(location, isDestroyed)

			end

	-               local newStructure = {"structure", location, {parts = structure}}
	+               local newStructure = {"structure", 0, {parts = structure}}
			table.insert(self.events.create, newStructure)
		end
	 end

Now we will run the game and go over that code path. We get the
following stack trace:

	Error: world/worldObjects.lua:7: attempt to index local 'location' (a number value)
	stack traceback:
		[string "boot.lua"]:777: in function '__index'
		world/worldObjects.lua:7: in function '__create'
		class.lua:5: in function 'parent'
		class.lua:4: in function 'objectClass'
		world/world.lua:210: in function 'update'
		gamestates/inGame.lua:177: in function <gamestates/inGame.lua:155>
		stackManager.lua:59: in function 'update'
		[string "boot.lua"]:612: in function <[string "boot.lua"]:594>
		[C]: in function 'xpcall'

There is nothing in this stack trace that mentions structure.lua. We
just know that someone inserted an invalid event into
world.events.create, and world.lua failed when it tried to process that
event. With the new system, this is the stack trace we get:

	Error: world/worldObjects.lua:7: attempt to index local 'location' (a number value)
	stack traceback:
		[string "boot.lua"]:777: in function '__index'
		world/worldObjects.lua:7: in function '__create'
		class.lua:5: in function 'parent'
		class.lua:4: in function 'Structure'
		world/structure.lua:347: in function 'fn'
		world/world.lua:206: in function 'update'
		gamestates/inGame.lua:177: in function <gamestates/inGame.lua:155>
		stackManager.lua:59: in function 'update'
		[string "boot.lua"]:612: in function <[string "boot.lua"]:594>
		[C]: in function 'xpcall'

Because it's the caller's responsibility to create the object, the
caller will fail. Now structure.lua is in the stack trace. "Attempt to
index local 'location' (a number value)" makes sense; we can see that
structure.lua is passing a number instead of a table to Structure() on
line 347:

	self.createObject(function(worldInfo)
		return (Structure(worldInfo, 0, {parts = structure}))
	end)