Improve missile accuracy
Use a pid controller to control the direction of missiles instead
of the simpler `+/- torque` method. The new controller takes into
consideration the current angular velocity of missiles and uses
variable torque instead of full blast to the left or to the right.
Adding these pieces give missiles more stability and accuracy.

Co-authored-by: randomguyeatspotato <randomguyeatspotato@gmail.com>