Simplify AppImage build script
Now that LOVE ships an official AppImage, it's a bit easier to build an
AppImage for Synthein. We just unpack LOVE's AppImage, drop in our .love
file, update the icon and .desktop file, and we're done. This is more
similar to how the Mac and Windows packages work: just download the
official LOVE build and add our stuff to it.

Additionally, the AppImage now ships with our Rust code, so it is able
to load the Rust-based modules.