Add an integration test (#350)
This is the first of many integration tests, where we test features of
the game while it is running via it's interface(s). The first test
starts the game in a scene. It should check to see if it fails and stop
the game once the tests are complete, but that's not yet implemented.