Add the first real integration test (#352)
This adds a test of the heal module that starts the game and watches an
injured player to see if it heals after some time. Currently this test
fails because the heal module appears to be broken.

In test-general1, I moved the repair part away from the player so that
it wouldn't repair the player. I only want to see the heal module
working.

In console.lua, I changed the read-evaluate-print-loop to only print if
there's output. This removes blank lines from the console output for
commands that return no output.