Convert heal.lua to Rust
heal.rs is quite a bit bigger than heal.rs, but most of the bulk and
complexity comes from being between two layers of Lua. It needs to call
hull.repair, which is a Lua function.

self.hull is a reference to the hull module that we need to hold onto
between calls into Rust code. All Lua objects have a lifetime of 'lua,
which is bound to the life of the Lua object that we receive in all
callbacks from Lua into Rust code. That Lua object only lives until the
end of the Rust callback, so we can't keep any Lua objects that live
longer than a function call.

mlua's Lua registry is a way around that. It's basically a global table
where we can store things and ask for them again later. self.hull is a
key that we can use to look up the actual hull module in the Lua
registry.