Print the problematic scene's name in test_all_scenes
This test was supposed to print the name of the scene that crashed, but
it had a bug that prevented it from doing that. run_test.sh runs all
tests in `set -e` mode, which means "if any command fails, abort the
script". This means that once LOVE exited with an error, the test would
abort without ever getting to the code where we print the scene name.

Now we catch and handle the error inside the test.