Make player 2's cursor visible (#404)
Before this change, we limited the cursor of player 2 to the right half
of the screen. For example, for an 800x600 screen, 400 ≤ x ≤ 800.
However, this doesn't work because we now render each player's viewport
to a canvas, where the top left corner is normalized to 0, 0.

Now, player's cursors are limited to 0 to the width and height of the
screen.