Remove redundant gamesave.lua (#405)
When SaveMenu was written, it copied the save file handling portion of
gamesave.lua. The remaining code was small enough that the file wasn't
needed.

Additionally, this moves file loading out of InGame. InGame.Load now
expects an iterator that it can pass to SceneParser.loadScene directly.
The input and output format that InGame understands is a serialized
world in the SceneParser format. It doesn't work with files directly
anymore.