Prevent the consolereader thread from hanging on quit
The infinite loop of the consolereader thread would continue forever
even after the main thread was done. This made it impossible to fully
quit the game without sending a KILL signal.

Now the reader thread will break out of the loop when the game exits,
but it will still wait for one last line of input. Pressing enter in the
terminal will end it.

I haven't found a way to wake up the thread from io.read() so that it
can notice that it should quit right away. Something clever you can do
with threads waiting on input from a socket is to just write to the
socket. But writing to stdin in Lua blocks, causing the whole game to
hang. Closing stdin might work, but Lua is hardcoded to prevent closing
stdin, stdout, and stderr (I checked the source code and found the
answer in liolib.c). So we'll just have to live with pressing enter one
more time for now.