Fix screen splitting for wider screens
The old algorithm for splitting a screen based on the number of players
had problems. It worked fine for a more square 4:3 aspect ratio screen,
but for widescreen ratios, it ran into problems after 3 players. For 4
players, it would add a third column to make a 3x2 grid.

On my DevTerm with its super wide 16:6 ratio, it ran into problems with
only a single player. For 1 player, it would split the screen into two
columns.

The new iterative algorithm only adds a new column or row when
necessary, and it tries to balance the columns and rows at the end to
make the grid sections as square as possible.