Clean up camera.lua's handling of game over
Previously, there were two paths that checked for the player ship being
destroyed in Camera:drawPlayer(). The ship.body:isDestroyed() path was
never used, because we don't destroy structure bodies with :destroy()
when the structure is destroyed. Now the indented functionality from
both paths is combined into one path that actually runs.

Also, don't try to set the camera's body to the player's ship body if
the ship is destroyed. That causes a crash.