Draw the selection cursor on the screen
Now the cursor appears on the screen, but the position is still wrong.
We need to do some of the translations that we do for world objects
(invert Y axis, rotate based on camera rotation, etc.) but not all (we
don't want to scale the cursor and circle menu to be tiny when you zoom
out).