Move selection translations to camera.lua
Now camera.lua knows about 3 kinds of transformations:

1. objects in the world
2. hud elements drawn to fixed positions on the screen
3. hud elements that point to a world object

The third type is new. It scales, rotates, translates so that the object
is always positioned over the world object it refers to, but it still
keeps the normal scale, ignoring zoom.

This last feature isn't quite done yet. I cheat it a bit by reversing
the effects of scaling manually on all love.draw calls inside the
drawLabels() function. This should be fixable in the future though.