Convert userdata connection message to a debug log (#443)
This message is useful, but it's also nice to be able to turn it off.
It's still on by default, because the ./run command starts the game in
debug mode. But it's possible to turn it off by leaving debug mode with
f12. This makes it easier to see other messages, such as some log:info
messages I'm using for something I'm working on.