Fix crash when particles are in the world
In 82302b3, I added an assumption that all world objects have a body
that has user data set on it. Most do, but particles previously did not
have a user data object set on their body, only a fixture attached to
the body. This caused an index into nil error that crashed the game
whenever particles were in the world, such as when a part is destroyed.

We want all bodies in the world to have user data for other reasons, so
the fix is to make sure particles have user data.

Unrelated: Also make the size of the particle fixture the right size. It
should be 2 blocks by 2 blocks, whereas it was previously 40 pixels by
40 pixels. This doesn't effect anything really, but it did cause weird
behavior when interacting with other bugs.