Prevent the unit tests from hanging (#453)
The unit test suite used to hang, waiting on console input. Just like
when you run the game normally you need to press enter once to stop the
console, the tests need a newline. This moves the console reader past
the line where it block waiting for input and allows it to check to see
if it should quit.