Write error messages to stderr
During tests, output is captured and stored in a temporary file that the
test code can inspect and make assertions about. We don't want error
messages and stack traces to go to the same place; that causes the
errors to get swallowed and deleted after the test. Errors should be
printed to the terminal after the test. The test runner already captures
and prints standard error, so printing to standard error has the
intended effect.