Fix crash in headless mode
The game was crashing in headless mode because of a comparison with a
stub in canvasUtils. canvasUtils.isWithin() gets the dimensions of a
canvas and then does comparisons on it. Since canvases are stubbed,
getDimensions returns a stub, which was not comparable until now.

Also, add the field `stub = true` to stubs to make it easier to notice
them when debugging.

I'll leave a marker.
Lost ones will see it and say:
Ah, this is a stub.