Move camera back when done
The last version didn't work because we were setting the camera target
to the player every frame, which reset the animation percent to zero
every frame.

Now we no longer set the camera target every frame. Like when we switch
to the building structure, we switch back on an event from
selection.lua.