Add smooth camera movement (#459)
When connecting a part to a structure, the camera moves to that structure
while doing the build action. Now that camera movement is linearly
interpolated.

I added a new Animation class. It could potentially support other
animation functions.

It's aware of x, y, and angle. It could be renamed cameraPan.lua, or
maybe in the future it will support multiple animations, one of which
is a pan. We'll see.