Go back to using ipairs
The `local times` line was necessary to avoid the high cost of fetching
drawTimeLogger.times from Rust and converting it to a Lua table.

However, if we go back to using ipairs and reverse the x axis
(chartWidth-x), it has the same effect, and we don't need an extra line
just for performance.