Upgrade mlua to 0.10.3
There were several breaking changes in 0.9 and 0.10. The biggest change
is that the 'lua lifetime is no longer required. This fixes some issues
I was having in another feature branch where one mlua API returned an
object with one lifetime and a different API accepted an object with a
different lifetime. Now neither of them has a lifetime specifier.