Fix camera spinning many rotations when shifting to a new target (#494)
If the current target of the camera has rotated a full circle many
times, and the new target has not, the pan animation would spin the
camera wildly to interpolate between the two angles.

This shouldn't happen because we normalize the camera angle between 0
and 2 * pi in Camera:update(). However, Camera:setTarget() accidentally
discarded that information by re-fetching the current angle from the
target's body.

By deleting this re-fetch and just using the already stored self.angle,
we can fix the bug.