AppImage: Switch to "fused" mode
When you run a game with LOVE, there are two ways to do it:

1. Run the love program, plus a .love file or directory: `love src` or
   `love synthein.love`. This is mainly used for development.
2. "Fuse" the game by packaging as a .love file and concatenating it to
   the love program.

There are several advantages to fusing the game to the love binary. The
biggest advantage is that the save directory is ~/.local/share/synthein
instead of ~/.local/share/love/synthein. It also makes starting the game
simpler. There's no need to tell love where to find the game. It knows
exactly where to find it because it's part of the same file.

This actually fixes a bug that was introduced when we switched to LOVE
11.4. It appears that something changed about how LOVE tries to find
.love files in that version, because it wasn't able to find
synthein.love. The AppImage build script has been broken since that
time.

I tried to use fused mode before, but I ran into a problem. I think LOVE
was only looking for syntheinrust.so outside the AppImage. In the
current version of LOVE, that problem seems to be fixed. It looks inside
the AppImage in exactly one place: lib/lua/5.1/syntheinrust.so.